+++ src/test/kotlin/org/ionproject/integration/extractor/TabulaPdfExtractorTests.kt
@@ -2,7 +2,7 @@ package org.ionproject.integration.extractor
 
 import java.io.File
 import kotlinx.coroutines.runBlocking
-import org.ionproject.integration.extractor.implementation.TabulaPdfExtractor
+import org.ionproject.integration.extractor.implementations.TabulaPdfExtractor
 import org.ionproject.integration.utils.Try
 import org.junit.jupiter.api.Assertions
 import org.junit.jupiter.api.Test
